Add files via upload
The tree will be structured as X(Y, Z). Where X is the parent of Y and Z. Y,Z branch down from X.

   1
   / \
  2   3
 /   / \
4   5   6

The lowest common ancestor is the lowest point of the tree where nodes X,Y have a connection.
For now we assume the tree is a binary tree.